refactor: Cleanup interface for spawn package
