+++ src/publisher.ts
@@ -2,8 +2,8 @@ import q = require('q');
 import path = require('path');
 import fs = require('fs');
 import logger from './logger';
-import spawn from './spawn/spawn';
-import { DEFAULT_ARG, SpawnArguments } from './spawn/arguments';
+import spawn from './spawn';
+import { DEFAULT_ARG, SpawnArguments } from './spawn';
 import { deprecate } from 'util';
 import pactStandalone from './pact-standalone';
 const checkTypes = require('check-types');
