+++ src/service.ts
@@ -6,8 +6,8 @@ import events = require('events');
 import http = require('request');
 import q = require('q');
 import logger from './logger';
-import spawn from './spawn/spawn';
-import { SpawnArguments } from './spawn/arguments';
+import spawn from './spawn';
+import { SpawnArguments } from './spawn';
 import { ChildProcess } from 'child_process';
 const mkdirp = require('mkdirp');
 const checkTypes = require('check-types');
