+++ src/spawn/index.ts
@@ -1,0 +1,2 @@
+export * from './arguments';
+export { default } from './spawn';
