+++ src/stub.ts
@@ -1,5 +1,5 @@
 // tslint:disable:no-string-literal
-import { DEFAULT_ARG } from './spawn/arguments';
+import { DEFAULT_ARG } from './spawn';
 import { AbstractService } from './service';
 import { deprecate } from 'util';
